ml/requirements.txt
# FastAPI and web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
httpx==0.28.1
scipy==1.15.2

# Google Generative AI
google-generativeai==0.8.3
google-genai==0.3.0

# Face recognition and computer vision
deepface==0.0.93
retina-face==0.0.17
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84

# NLP and text processing
nltk==3.9.1

# Document processing
pdfplumber==0.11.4
python-docx==1.1.2
pdf2image==1.17.0
pytesseract==0.3.13

# Data processing
numpy==1.26.4
pandas==2.2.3

# Streamlit (used in face_matching_helper)
streamlit==1.41.1

# Additional dependencies
pydantic==2.10.5
IPython==8.31.0

reportlab
pypdf
requests
